setting up block header to be reused in chain
