+++ ironfish/src/rpc/index.ts
@@ -3,8 +3,9 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 export * from './adapters'
 export * from './clients'
+export * from './messageBuffer'
 export * from './response'
 export * from './routes'
 export * from './server'
 export * from './stream'
-export * from './messageBuffer'
+export * from './types'
