+++ ironfish/src/rpc/routes/event/index.ts
@@ -5,4 +5,3 @@
 export * from './onGossip'
 export * from './onReorganizeChain'
 export * from './onTransactionGossip'
-export { RpcBlockHeader } from './types'
