+++ ironfish/src/rpc/routes/event/onGossip.ts
@@ -5,8 +5,8 @@ import * as yup from 'yup'
 import { Assert } from '../../../assert'
 import { FullNode } from '../../../node'
 import { BlockHeader } from '../../../primitives'
+import { RpcBlockHeader, RpcBlockHeaderSchema, serializeRpcBlockHeader } from '../../types'
 import { ApiNamespace, routes } from '../router'
-import { RpcBlockHeader, RpcBlockHeaderSchema, serializeRpcBlockHeader } from './types'
 
 export type OnGossipRequest = undefined
 export type OnGossipResponse = { blockHeader: RpcBlockHeader }
